panel: Add explicit handling of application state
When the active application is changed, the application will receive
an application state change event. Commonly this does not necessarily
also change the focus state of the application (and indeed the
application state change is not forwarded into the widget hierarchy
by Qt).

Thus when the application is activated again, the browser panel does
not regain focus (and keyboard focus) even if it had that focus before
the application was deactivated.

While CEF does receive an activation event on macOS (because it
subscribes to corresponding AppKit event notifications), the same is
not true on Windows.

To ensure that the browser panel indeed regains focus, the state is
thus set explicitly corresponding to the application event states.