panel: Propagate "focus-in" event to browser host
The CEF instance is wrapped in another QWindow instance for Windows
and Linux and thus consumes any focus events that are propagated through
the application.

To ensure that a wrapped CEF instance regains focus after its wrapping
window receives focus, the event needs to be explicitly sent to the
browser host instance it wraps.